fixed-a-bug-and-added-middleware

+++ routes/web.php
@@ -32,28 +32,28 @@ Route::post('/home', [ProjectController::class, 'create'])->name('create_project
 Route::prefix('/home')->group(function(){
 
     //All projects
-    Route::get('/projects', [UserProjectController::class, 'getAllProjects'])->name('allprojects');
+    Route::get('/projects', [UserProjectController::class, 'getAllProjects'])->name('allprojects')->middleware('auth');
 
     //Show Projects
-    Route::get('/project/{id}', [ProjectController::class, 'show_project'])->name('show_project');
+    Route::get('/project/{id}', [ProjectController::class, 'show_project'])->name('show_project')->middleware('auth');
 
     //Add task to  projects
-    Route::post('/project/{id}/task', [TaskController::class, 'create'])->name('add_task');
+    Route::post('/project/{id}/task', [TaskController::class, 'create'])->name('add_task')->middleware('auth');
 
     //get tasks 
-    Route::get('/tasks', [TaskController::class, 'getAll'])->name('alltasks');
+    Route::get('/tasks', [TaskController::class, 'getAll'])->name('alltasks')->middleware('auth');
 
     //Edit Task
-    Route::get('/tasks/{id}/edit', [TaskController::class, 'editTask'])->name('edit_task');
+    Route::get('/tasks/{id}/edit', [TaskController::class, 'editTask'])->name('edit_task')->middleware('auth');
 
     //Update Task
-    Route::put('/tasks/{id}/update', [TaskController::class, 'update'])->name('update_task');
+    Route::put('/tasks/{id}/update', [TaskController::class, 'update'])->name('update_task')->middleware('auth');
 
     //Delete Task
-    Route::delete('/tasks/{id}/delete', [TaskController::class, 'delete'])->name('delete_task');
+    Route::delete('/tasks/{id}/delete', [TaskController::class, 'delete'])->name('delete_task')->middleware('auth');
 
     //get tasks from projects dropdown
-    Route::post('/{id}', [TasksComposer::class, 'tt'])->name('getdata');
+    Route::post('/{id}', [TasksComposer::class, 'tt'])->name('getdata')->middleware('auth');
 
 
 });
